requirements.txt
numpy
scipy
pandas
pyyaml
matplotlib
pytest
healpy
pymaster
